src/analysis/solver/nonlinear.ts
import { SolverState } from '../../../spec/Solver';
import { Matrix, createMatrix } from './matrix';

export function solveNonlinear(state: SolverState): SolverState {
  // This is a placeholder for nonlinear optimization
  // In a real implementation, this would:
  // 1. Use gradient descent or other nonlinear methods
  // 2. Handle nonlinear constraints
  // 3. Find local/global optima
  return state;
}

export function calculateGradient(
  state: SolverState,
  point: number[],
  epsilon: number = 1e-6
): number[] {
  const gradient: number[] = [];
  const n = point.length;
  
  // Calculate partial derivatives
  for (let i = 0; i < n; i++) {
    const h = Math.max(Math.abs(point[i] * epsilon), epsilon);
    const forward = [...point];
    const backward = [...point];
    
    forward[i] += h;
    backward[i] -= h;
    
    // Central difference approximation
    gradient[i] = (evaluateObjective(state, forward) - evaluateObjective(state, backward)) / (2 * h);
  }
  
  return gradient;
}

function evaluateObjective(state: SolverState, point: number[]): number {
  // This is a placeholder
  // In a real implementation, this would evaluate the objective function
  return 0;
}